+++ axum-extra/src/response/mod.rs
@@ -9,10 +9,17 @@ mod attachment;
 #[cfg(feature = "multipart")]
 pub mod multiple;
 
-#[cfg(feature = "error_response")]
+#[cfg(feature = "error-response")]
 mod error_response;
 
-#[cfg(feature = "error_response")]
+#[cfg(feature = "file-stream")]
+/// Module for handling file streams.
+pub mod file_stream;
+
+#[cfg(feature = "file-stream")]
+pub use file_stream::FileStream;
+
+#[cfg(feature = "error-response")]
 pub use error_response::InternalServerError;
 
 #[cfg(feature = "erased-json")]