+++ axum-extra/CHANGELOG.md
@@ -11,13 +11,20 @@ and this project adheres to [Semantic Versioning].
   request in many cases; see its documentation for details ([#2475])
 - **changed:** Deprecated `OptionalPath<T>` and `OptionalQuery<T>` ([#2475])
 - **fixed:** `Host` extractor includes port number when parsing authority ([#2242])
+- **changed:** The `multipart` feature is no longer on by default ([#3058])
 - **added:** Add `RouterExt::typed_connect` ([#2961])
 - **added:** Add `json!` for easy construction of JSON responses ([#2962])
+- **added:** Add `InternalServerError` response for logging an internal error
+  and returning HTTP 500 in a convenient way. ([#3010])
+- **added:** Add `FileStream` for easy construction of file stream responses ([#3047])
 
 [#2242]: https://github.com/tokio-rs/axum/pull/2242
 [#2475]: https://github.com/tokio-rs/axum/pull/2475
+[#3058]: https://github.com/tokio-rs/axum/pull/3058
 [#2961]: https://github.com/tokio-rs/axum/pull/2961
 [#2962]: https://github.com/tokio-rs/axum/pull/2962
+[#3010]: https://github.com/tokio-rs/axum/pull/3010
+[#3047]: https://github.com/tokio-rs/axum/pull/3047
 
 # 0.10.0